+++ lib/active_function/base.rb
@@ -4,12 +4,12 @@ module ActiveFunction
   class Base
     require "active_function/functions/core"
     require "active_function/functions/callbacks"
-    require "active_function/functions/params"
+    require "active_function/functions/strong_parameters"
     require "active_function/functions/rendering"
 
     include Functions::Core
     include Functions::Rendering
-    include Functions::Params
+    include Functions::StrongParameters
     include Functions::Callbacks
   end
 end