+++ test/support/active_function_helper.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module ActiveFunctionHelper
+end
+
+Minitest::Test.include ActiveFunctionHelper