+++ test/test_helper.rb
@@ -1,6 +1,13 @@
 # frozen_string_literal: true
 
-$LOAD_PATH.unshift File.expand_path("../lib", __dir__)
-require "active_function/functions"
+require "bundler/setup"
+require "ruby-next/language/runtime" unless ENV["CI"]
 
 require "minitest/autorun"
+require "minitest/reporters"
+
+require "active_function"
+
+require "./test/support/active_function_helper"
+
+Minitest::Reporters.use! [Minitest::Reporters::SpecReporter.new(color: true)]