+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 
-## [0.0.1] - 2022-07-29
+## [0.1.0] - 2022-12-28
 
-- Initial release
+- Add `before_action` and `after_action` callbacks with `:only` and `:if` options
+- Add `ActiveFunction::Base#params` with `require`, `permit`, `[]` and `to_h` public methods
+- Add specs for `ActiveFunction::Base.params` module
+- Add `ActiveFunction::Base#render` with `:status`, `:json` and `:head` options
+- Add RBS signature
+- Add RBS Aws Event Types
+- Add public interface type `steep` checking
+- Add `ruby-next` transpiler
+-